refactor(core): unify chart rendering into @silurus/ooxml-core
Move duplicated chart rendering from pptx and xlsx into a single shared
implementation in core. Both packages now delegate to `renderChart` from
@silurus/ooxml-core and drop ~1,400 lines of near-duplicate code.

New in core:
- `ChartModel` / `ChartSeries` canonical types (ChartType embeds direction
  and grouping: clusteredBar, stackedBarHPct, stackedLinePct, …)
- `renderChart(ctx, model, rect)` — unified renderer covering bar, line,
  area, pie / doughnut, radar, scatter, and waterfall families.
- Supports valMin / valMax, plotAreaBg, dataPointColors (pie slices),
  subtotalIndices (waterfall), catAxisHidden / valAxisHidden.

Consumers:
- pptx: ChartElement keeps its Rust-flat shape; renderer passes fields to
  `renderChart` via an inline adapter.
- xlsx: `ChartData` remains the parser output type; `adaptChartData`
  normalizes barDir+grouping into canonical ChartModel.chartType before
  delegating.

No WASM or Rust changes — normalization lives in the TS adapter layer so
parser contracts stay untouched. Type checking passes for core, pptx,
xlsx, and docx.
