+++ packages/core/src/index.ts
@@ -1,1 +1,3 @@
 export * from './types/common';
+export * from './types/chart';
+export { renderChart } from './chart/renderer';
