+++ packages/pptx/src/renderer.ts
@@ -4,8 +4,6 @@ import type {
   ShapeElement,
   PictureElement,
   TableElement,
-  ChartElement,
-  ChartSeriesData,
   Fill,
   Stroke,
   TextBody,
@@ -16,6 +14,7 @@ import type {
   RenderOptions,
   TabStop,
 } from './types';
+import { renderChart } from '@silurus/ooxml-core';
 
 /** EMU per point (OOXML: 1 pt = 12700 EMU). Used to scale font sizes with the canvas. */
 const PT_TO_EMU = 12700;
@@ -2730,718 +2729,9 @@ function applyStroke(ctx: CanvasRenderingContext2D, stroke: Stroke | null, scale
 }
 
 // ─── Chart rendering ────────────────────────────────────────────────────────
+// Chart rendering is delegated to @silurus/ooxml-core's unified renderer.
+// See renderChart(ctx, el, scale) below.
 
-function niceStep(range: number, targetSteps = 5): number {
-  const raw = range / targetSteps;
-  const mag = Math.pow(10, Math.floor(Math.log10(raw)));
-  const normed = raw / mag;
-  const nice = normed < 1.5 ? 1 : normed < 3.5 ? 2 : normed < 7.5 ? 5 : 10;
-  return nice * mag;
-}
-
-function renderStackedBarChart(
-  ctx: CanvasRenderingContext2D, el: ChartElement, scale: number
-) {
-  const x = emuToPx(el.x, scale);
-  const y = emuToPx(el.y, scale);
-  const w = emuToPx(el.width, scale);
-  const h = emuToPx(el.height, scale);
-
-  const hasTitle = el.title && el.title.trim().length > 0;
-  const titleH = hasTitle ? h * 0.09 : 0;
-  const padL = w * 0.11;
-  const padR = w * 0.04;
-  const padT = h * 0.06 + titleH;
-  const padB = h * 0.13;
-  const px0 = x + padL;
-  const py0 = y + padT;
-  const pw  = w - padL - padR;
-  const ph  = h - padT - padB;
-
-  const cats = el.categories;
-  const n = cats.length;
-  if (n === 0) return;
-
-  // Compute stacked totals to find axis scale
-  const totals = Array.from({ length: n }, (_, i) =>
-    el.series.reduce((s, ser) => s + (ser.values[i] ?? 0), 0)
-  );
-  const dataMax = el.valMax ?? Math.max(...totals) * 1.05;
-  if (dataMax <= 0) return;
-
-  ctx.save();
-
-  // Chart title
-  if (hasTitle) {
-    ctx.font = `bold ${Math.round(h * 0.062)}px sans-serif`;
-    ctx.fillStyle = '#222';
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'middle';
-    ctx.fillText(el.title!, x + w / 2, y + titleH / 2 + h * 0.015);
-  }
-
-  // Grid lines + Y-axis labels
-  const step = niceStep(dataMax);
-  const labelFontSize = Math.round(h * 0.045);
-  ctx.font = `${labelFontSize}px sans-serif`;
-  ctx.textAlign = 'right';
-  ctx.textBaseline = 'middle';
-  ctx.fillStyle = '#666';
-  ctx.strokeStyle = '#e8e8e8';
-  ctx.lineWidth = 0.7;
-  for (let v = 0; v <= dataMax + step * 0.5; v += step) {
-    const gy = py0 + ph * (1 - v / dataMax);
-    ctx.beginPath();
-    ctx.moveTo(px0, gy);
-    ctx.lineTo(px0 + pw, gy);
-    ctx.stroke();
-    ctx.fillText(v.toLocaleString(), px0 - 4, gy);
-  }
-
-  // Y-axis line and X-axis baseline
-  ctx.strokeStyle = '#bbb';
-  ctx.lineWidth = 1;
-  ctx.beginPath();
-  ctx.moveTo(px0, py0);
-  ctx.lineTo(px0, py0 + ph);
-  ctx.lineTo(px0 + pw, py0 + ph);
-  ctx.stroke();
-
-  // Bars
-  const barW = (pw / n) * 0.55;
-  const gapW = pw / n;
-
-  const labelFontSizeBar = Math.round(h * 0.038);
-
-  el.series.forEach((ser, si) => {
-    ctx.fillStyle = ser.color ? `#${ser.color}` : `hsl(${210 + si * 40}, 60%, ${50 - si * 8}%)`;
-    for (let i = 0; i < n; i++) {
-      const v = ser.values[i] ?? 0;
-      if (v === 0) continue;
-      const base = el.series.slice(0, si).reduce((s, ps) => s + (ps.values[i] ?? 0), 0);
-      const barH = ph * (v / dataMax);
-      const bx = px0 + gapW * i + (gapW - barW) / 2;
-      const by = py0 + ph * (1 - (base + v) / dataMax);
-      ctx.fillRect(bx, by, barW, barH);
-
-      // Data labels: draw value centered in bar segment if it's tall enough
-      if (el.showDataLabels && barH >= labelFontSizeBar * 1.2) {
-        ctx.save();
-        ctx.font = `${labelFontSizeBar}px sans-serif`;
-        ctx.fillStyle = '#ffffff';
-        ctx.textAlign = 'center';
-        ctx.textBaseline = 'middle';
-        const labelText = v.toLocaleString();
-        const labelX = bx + barW / 2;
-        const labelY = by + barH / 2;
-        ctx.fillText(labelText, labelX, labelY);
-        ctx.restore();
-      }
-    }
-  });
-
-  // X-axis category labels
-  ctx.textAlign = 'center';
-  ctx.textBaseline = 'top';
-  ctx.fillStyle = '#666';
-  ctx.font = `${Math.round(h * 0.042)}px sans-serif`;
-  for (let i = 0; i < n; i++) {
-    const cx = px0 + gapW * i + gapW / 2;
-    ctx.fillText(cats[i], cx, py0 + ph + 4);
-  }
-
-  // Legend (top, below title)
-  if (el.series.length > 1) {
-    const legY = y + titleH + h * 0.03;
-    const legFontSize = Math.round(h * 0.042);
-    ctx.font = `${legFontSize}px sans-serif`;
-    ctx.textAlign = 'left';
-    ctx.textBaseline = 'middle';
-    let legX = px0;
-    el.series.forEach((ser, si) => {
-      const col = ser.color ? `#${ser.color}` : `hsl(${210 + si * 40}, 60%, ${50 - si * 8}%)`;
-      ctx.fillStyle = col;
-      ctx.fillRect(legX, legY - legFontSize / 2, legFontSize, legFontSize);
-      ctx.fillStyle = '#444';
-      ctx.fillText(ser.name, legX + legFontSize + 4, legY);
-      legX += legFontSize + 4 + ctx.measureText(ser.name).width + 16;
-    });
-  }
-
-  ctx.restore();
-}
-
-function renderStackedBarHChart(
-  ctx: CanvasRenderingContext2D, el: ChartElement, scale: number
-) {
-  const x = emuToPx(el.x, scale);
-  const y = emuToPx(el.y, scale);
-  const w = emuToPx(el.width, scale);
-  const h = emuToPx(el.height, scale);
-
-  // When the value axis is hidden, reclaim its bottom padding to give bars more room.
-  const padL = el.catAxisHidden ? w * 0.04 : w * 0.22;  // category labels on left
-  const padR = w * 0.04;
-  const padT = h * 0.12;  // legend + top labels
-  const padB = el.valAxisHidden ? h * 0.02 : h * 0.10;  // x-axis labels
-  const px0 = x + padL;
-  const py0 = y + padT;
-  const pw  = w - padL - padR;
-  const ph  = h - padT - padB;
-
-  const cats = el.categories;
-  const n = cats.length;
-  if (n === 0) return;
-
-  const totals = Array.from({ length: n }, (_, i) =>
-    el.series.reduce((s, ser) => s + (ser.values[i] ?? 0), 0)
-  );
-  const dataMax = el.valMax ?? Math.max(...totals) * 1.05;
-  if (dataMax <= 0) return;
-
-  ctx.save();
-
-  // Grid lines + X-axis (value) labels — both follow the value-axis visibility flag,
-  // since gridlines are anchored to the value axis ticks.
-  if (!el.valAxisHidden) {
-    const step = niceStep(dataMax);
-    ctx.font = `${Math.round(h * 0.045)}px sans-serif`;
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'top';
-    ctx.fillStyle = '#666';
-    ctx.strokeStyle = '#e8e8e8';
-    ctx.lineWidth = 0.7;
-    for (let v = 0; v <= dataMax + step * 0.5; v += step) {
-      const gx = px0 + pw * (v / dataMax);
-      ctx.beginPath(); ctx.moveTo(gx, py0); ctx.lineTo(gx, py0 + ph); ctx.stroke();
-      ctx.fillText(v.toLocaleString(), gx, py0 + ph + 4);
-    }
-
-    // X-axis line and Y-axis baseline
-    ctx.strokeStyle = '#bbb';
-    ctx.lineWidth = 1;
-    ctx.beginPath();
-    ctx.moveTo(px0, py0);
-    ctx.lineTo(px0, py0 + ph);
-    ctx.lineTo(px0 + pw, py0 + ph);
-    ctx.stroke();
-  }
-
-  // Bars
-  const barH = (ph / n) * 0.55;
-  const gapH = ph / n;
-
-  const labelFontSizeH = Math.round(h * 0.038);
-
-  // PowerPoint horizontal bar charts display categories bottom-to-top in data order,
-  // so visually the first category (index 0) is at the BOTTOM and the last is at the TOP.
-  // We render in reverse visual order: data index i → visual row (n - 1 - i) from the top.
-  el.series.forEach((ser, si) => {
-    ctx.fillStyle = ser.color ? `#${ser.color}` : `hsl(${210 + si * 40}, 60%, ${50 - si * 8}%)`;
-    for (let i = 0; i < n; i++) {
-      const ri = n - 1 - i; // reversed visual row index
-      const v = ser.values[i] ?? 0;
-      if (v === 0) continue;
-      const base = el.series.slice(0, si).reduce((s, ps) => s + (ps.values[i] ?? 0), 0);
-      const segW = pw * (v / dataMax);
-      const bx = px0 + pw * (base / dataMax);
-      const by = py0 + gapH * ri + (gapH - barH) / 2;
-      ctx.fillRect(bx, by, segW, barH);
-
-      // Data labels: draw value centered in bar segment if wide enough
-      if (el.showDataLabels && segW >= labelFontSizeH * 2) {
-        ctx.save();
-        ctx.font = `${labelFontSizeH}px sans-serif`;
-        ctx.fillStyle = '#ffffff';
-        ctx.textAlign = 'center';
-        ctx.textBaseline = 'middle';
-        const labelText = v.toLocaleString();
-        const labelX = bx + segW / 2;
-        const labelY = by + barH / 2;
-        ctx.fillText(labelText, labelX, labelY);
-        ctx.restore();
-      }
-    }
-  });
-
-  // Y-axis (category) labels — rendered in reverse: cat[n-1] at top, cat[0] at bottom
-  if (!el.catAxisHidden) {
-    ctx.textAlign = 'right';
-    ctx.textBaseline = 'middle';
-    ctx.fillStyle = '#444';
-    ctx.font = `${Math.round(h * 0.042)}px sans-serif`;
-    for (let i = 0; i < n; i++) {
-      const ri = n - 1 - i; // reversed visual row index
-      const cy = py0 + gapH * ri + gapH / 2;
-      ctx.fillText(cats[i], px0 - 6, cy);
-    }
-  }
-
-  // Legend (top)
-  if (el.series.length > 1) {
-    const legY = y + padT * 0.4;
-    const legFontSize = Math.round(h * 0.042);
-    ctx.font = `${legFontSize}px sans-serif`;
-    ctx.textAlign = 'left';
-    ctx.textBaseline = 'middle';
-    let legX = px0;
-    el.series.forEach((ser, si) => {
-      const col = ser.color ? `#${ser.color}` : `hsl(${210 + si * 40}, 60%, ${50 - si * 8}%)`;
-      ctx.fillStyle = col;
-      ctx.fillRect(legX, legY - legFontSize / 2, legFontSize, legFontSize);
-      ctx.fillStyle = '#444';
-      ctx.fillText(ser.name, legX + legFontSize + 4, legY);
-      legX += legFontSize + 4 + ctx.measureText(ser.name).width + 16;
-    });
-  }
-
-  ctx.restore();
-}
-
-function renderWaterfallChart(
-  ctx: CanvasRenderingContext2D, el: ChartElement, scale: number
-) {
-  const x = emuToPx(el.x, scale);
-  const y = emuToPx(el.y, scale);
-  const w = emuToPx(el.width, scale);
-  const h = emuToPx(el.height, scale);
-
-  const padL = w * 0.11;
-  const padR = w * 0.04;
-  const padT = h * 0.08;
-  const padB = h * 0.18;
-  const px0 = x + padL;
-  const py0 = y + padT;
-  const pw  = w - padL - padR;
-  const ph  = h - padT - padB;
-
-  const vals = el.series[0]?.values ?? [];
-  const cats = el.categories;
-  const n = cats.length;
-  if (n === 0) return;
-
-  const subSet = new Set(el.subtotalIndices);
-
-  // Build bar segments: {start, end, isSub, isPos}
-  let running = 0;
-  const bars: Array<{ start: number; end: number; isSub: boolean; isPos: boolean }> = [];
-  for (let i = 0; i < n; i++) {
-    const v = vals[i] ?? 0;
-    const isSub = i === 0 || subSet.has(i);
-    if (isSub) {
-      bars.push({ start: 0, end: v, isSub: true, isPos: true });
-      running = v;
-    } else {
-      const start = v >= 0 ? running : running + v;
-      const end   = v >= 0 ? running + v : running;
-      bars.push({ start, end, isSub: false, isPos: v >= 0 });
-      running += v;
-    }
-  }
-
-  // Axis range
-  const allEnds = bars.map(b => b.end);
-  const allStarts = bars.map(b => b.start);
-  const rawMax = Math.max(...allEnds, ...allStarts);
-  const rawMin = Math.min(...allStarts, 0);
-  const dataRange = rawMax - rawMin;
-  if (dataRange <= 0) return;
-  const padded = dataRange * 1.1;
-  const dataMin = rawMin - dataRange * 0.05;
-  const dataMax = dataMin + padded;
-
-  // Grid
-  const step = niceStep(padded);
-  ctx.save();
-  const fontSize = Math.round(h * 0.042);
-  ctx.font = `${fontSize}px sans-serif`;
-  ctx.strokeStyle = '#e8e8e8';
-  ctx.lineWidth = 0.7;
-  ctx.fillStyle = '#666';
-  ctx.textAlign = 'right';
-  ctx.textBaseline = 'middle';
-  for (let v = Math.ceil(dataMin / step) * step; v <= dataMax; v += step) {
-    const gy = py0 + ph * (1 - (v - dataMin) / padded);
-    ctx.beginPath(); ctx.moveTo(px0, gy); ctx.lineTo(px0 + pw, gy); ctx.stroke();
-    ctx.fillText(v.toLocaleString(), px0 - 4, gy);
-  }
-
-  // Y-axis line and X-axis baseline (at dataMin level = y-axis origin)
-  ctx.strokeStyle = '#bbb';
-  ctx.lineWidth = 1;
-  ctx.beginPath();
-  ctx.moveTo(px0, py0);
-  ctx.lineTo(px0, py0 + ph);
-  ctx.lineTo(px0 + pw, py0 + ph);
-  ctx.stroke();
-
-  // Colors
-  const colorSub = '#196ECA';
-  const colorPos = '#5BA4E6';
-  const colorNeg = '#E46970';
-
-  const barW = (pw / n) * 0.55;
-  const gapW = pw / n;
-
-  bars.forEach((bar, i) => {
-    const bx = px0 + gapW * i + (gapW - barW) / 2;
-    const yTop = py0 + ph * (1 - (bar.end - dataMin) / padded);
-    const yBot = py0 + ph * (1 - (bar.start - dataMin) / padded);
-    const bh = Math.max(1, yBot - yTop);
-
-    if (bar.isSub) {
-      ctx.fillStyle = colorSub;
-      ctx.fillRect(bx, yTop, barW, bh);
-    } else {
-      // Delta bars: outlined (hollow) style
-      ctx.strokeStyle = bar.isPos ? colorPos : colorNeg;
-      ctx.lineWidth = 1.5;
-      ctx.strokeRect(bx + 0.75, yTop + 0.75, barW - 1.5, bh - 1.5);
-    }
-
-    // Connector line to next bar
-    if (i < n - 1) {
-      const nextBx = px0 + gapW * (i + 1) + (gapW - barW) / 2;
-      const connY = bar.isPos ? yTop : yBot;
-      ctx.strokeStyle = '#ccc';
-      ctx.lineWidth = 0.8;
-      ctx.setLineDash([3, 3]);
-      ctx.beginPath();
-      ctx.moveTo(bx + barW, connY);
-      ctx.lineTo(nextBx, connY);
-      ctx.stroke();
-      ctx.setLineDash([]);
-    }
-
-    // Value label above each bar
-    const rawVal = vals[i] ?? 0;
-    const labelText = rawVal < 0
-      ? `△ ${Math.abs(rawVal).toLocaleString()}`
-      : rawVal.toLocaleString();
-    ctx.fillStyle = '#595959';
-    ctx.font = `bold ${Math.round(h * 0.044)}px sans-serif`;
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'bottom';
-    ctx.fillText(labelText, bx + barW / 2, yTop - 3);
-  });
-
-  // X-axis labels (may contain newlines → split)
-  ctx.textAlign = 'center';
-  ctx.textBaseline = 'top';
-  ctx.fillStyle = '#666';
-  ctx.font = `${Math.round(h * 0.038)}px sans-serif`;
-  const labelY = py0 + ph + 4;
-  for (let i = 0; i < n; i++) {
-    const cx = px0 + gapW * i + gapW / 2;
-    const lines = cats[i].split(/\s+/);
-    lines.forEach((line, li) => ctx.fillText(line, cx, labelY + li * (fontSize + 2)));
-  }
-
-  ctx.restore();
-}
-
-// ─── Default color palette for charts without explicit series colors ─────────
-const CHART_PALETTE = [
-  '4472C4','ED7D31','A9D18E','FF0000','70AD47','4BACC6',
-  'FFC000','9E480E','843C0C','636363','255E91','967300',
-];
-function chartColor(idx: number, explicit: string | null | undefined): string {
-  return explicit ? `#${explicit}` : `#${CHART_PALETTE[idx % CHART_PALETTE.length]}`;
-}
-
-// ─── Line chart ──────────────────────────────────────────────────────────────
-function niceStepLine(range: number): number {
-  if (range <= 0) return 1;
-  const exp = Math.floor(Math.log10(range));
-  const frac = range / Math.pow(10, exp);
-  const step = frac < 1.5 ? 0.2 : frac < 3 ? 0.5 : frac < 7 ? 1 : 2;
-  return step * Math.pow(10, exp);
-}
-
-function renderLineChart(ctx: CanvasRenderingContext2D, el: ChartElement, scale: number) {
-  const x0 = emuToPx(el.x, scale), y0 = emuToPx(el.y, scale);
-  const w  = emuToPx(el.width, scale), h = emuToPx(el.height, scale);
-
-  const allVals = el.series.flatMap(s => s.values).filter((v): v is number => v != null);
-  const rawMax = el.valMax ?? (allVals.length ? Math.max(...allVals) : 1);
-  const rawMin = el.valMin ?? (allVals.length ? Math.min(...allVals) : 0);
-
-  const titleH = el.title ? Math.max(16, h * 0.09) : 0;
-  const axisFontSz = Math.max(8, Math.min(11, h * 0.05));
-  const yAxisLabelW = el.valAxisHidden ? 4 : axisFontSz * 2.5 + 6;
-  const catLabelH = el.catAxisHidden ? 2 : axisFontSz + 6;
-  const plotX = x0 + yAxisLabelW;
-  const plotY = y0 + titleH + 10;
-  const plotW = w - yAxisLabelW - w * 0.03;
-  const plotH = h - titleH - catLabelH - 8;
-  if (plotW <= 0 || plotH <= 0) return;
-
-  // Axis tick computation: use nice step when user didn't specify valMin/valMax
-  let axMin = rawMin, axMax = rawMax;
-  if (el.valMin == null || el.valMax == null) {
-    const step = niceStepLine(Math.max(1e-9, rawMax - rawMin));
-    axMin = Math.floor(rawMin / step) * step;
-    axMax = Math.ceil(rawMax / step) * step;
-  }
-  if (axMax === axMin) axMax = axMin + 1;
-  const tickStep = niceStepLine(axMax - axMin);
-  const range = axMax - axMin;
-
-  const toX = (i: number) => plotX + (el.categories.length > 1 ? (i / (el.categories.length - 1)) * plotW : plotW / 2);
-  const toY = (v: number) => plotY + plotH - ((v - axMin) / range) * plotH;
-
-  ctx.save();
-
-  // Plot area background
-  if (el.plotAreaBg) {
-    ctx.fillStyle = `#${el.plotAreaBg}`;
-    ctx.fillRect(plotX, plotY, plotW, plotH);
-  }
-
-  // Gridlines + Y tick labels
-  if (!el.valAxisHidden) {
-    ctx.font = `${axisFontSz}px sans-serif`;
-    ctx.textBaseline = 'middle';
-    const steps = Math.max(1, Math.round(range / tickStep));
-    for (let s = 0; s <= steps; s++) {
-      const v = axMin + s * tickStep;
-      const gy = toY(v);
-      ctx.strokeStyle = '#D8E2D0';
-      ctx.lineWidth = 0.75;
-      ctx.setLineDash([]);
-      ctx.beginPath(); ctx.moveTo(plotX, gy); ctx.lineTo(plotX + plotW, gy); ctx.stroke();
-      ctx.fillStyle = '#555'; ctx.textAlign = 'right';
-      const label = Number.isInteger(v) ? String(v) : v.toFixed(1);
-      ctx.fillText(label, plotX - 4, gy);
-    }
-  }
-
-  // Category labels
-  if (!el.catAxisHidden) {
-    ctx.fillStyle = '#555'; ctx.textBaseline = 'top'; ctx.textAlign = 'center';
-    ctx.font = `${axisFontSz}px sans-serif`;
-    const lblStep = el.categories.length > 8 ? Math.ceil(el.categories.length / 8) : 1;
-    for (let i = 0; i < el.categories.length; i += lblStep) {
-      ctx.fillText(el.categories[i], toX(i), plotY + plotH + 3);
-    }
-  }
-
-  // Series
-  const markerR = Math.max(2, h * 0.012);
-  for (let si = 0; si < el.series.length; si++) {
-    const ser = el.series[si];
-    const color = chartColor(si, ser.color);
-    ctx.strokeStyle = color; ctx.lineWidth = Math.max(1.5, h * 0.004); ctx.setLineDash([]);
-    ctx.beginPath();
-    let first = true;
-    for (let i = 0; i < ser.values.length; i++) {
-      const v = ser.values[i];
-      if (v == null) { first = true; continue; }
-      if (first) { ctx.moveTo(toX(i), toY(v)); first = false; } else { ctx.lineTo(toX(i), toY(v)); }
-    }
-    ctx.stroke();
-    // Markers + data labels
-    ctx.fillStyle = color;
-    for (let i = 0; i < ser.values.length; i++) {
-      const v = ser.values[i]; if (v == null) continue;
-      const px = toX(i), py = toY(v);
-      ctx.beginPath(); ctx.arc(px, py, markerR, 0, Math.PI * 2); ctx.fill();
-    }
-    if (el.showDataLabels) {
-      ctx.font = `${axisFontSz}px sans-serif`;
-      ctx.fillStyle = '#333'; ctx.textAlign = 'left'; ctx.textBaseline = 'middle';
-      for (let i = 0; i < ser.values.length; i++) {
-        const v = ser.values[i]; if (v == null) continue;
-        const label = Number.isInteger(v) ? String(v) : v.toFixed(1);
-        ctx.fillText(label, toX(i) + markerR + 2, toY(v));
-      }
-    }
-  }
-
-  // Title
-  if (el.title) {
-    ctx.fillStyle = '#333'; ctx.textAlign = 'center'; ctx.textBaseline = 'top';
-    ctx.font = `bold ${Math.max(10, h * 0.075)}px sans-serif`;
-    ctx.fillText(el.title, x0 + w / 2, y0 + 2);
-  }
-  ctx.restore();
-}
-
-// ─── Area chart ──────────────────────────────────────────────────────────────
-function renderAreaChart(ctx: CanvasRenderingContext2D, el: ChartElement, scale: number) {
-  const x0 = emuToPx(el.x, scale), y0 = emuToPx(el.y, scale);
-  const w  = emuToPx(el.width, scale), h = emuToPx(el.height, scale);
-
-  const allVals = el.series.flatMap(s => s.values).filter((v): v is number => v != null);
-  const dataMax = el.valMax ?? (allVals.length ? Math.max(...allVals) : 1);
-  const dataMin = Math.min(0, ...(allVals.length ? allVals : [0]));
-
-  const LABEL_H = h * 0.12, MARGIN = w * 0.05;
-  const plotX = x0 + MARGIN, plotW = w - MARGIN * 2;
-  const plotY = y0 + LABEL_H * 0.5, plotH = h - LABEL_H * 1.5;
-  const range = dataMax - dataMin || 1;
-
-  const toX = (i: number) => plotX + (el.categories.length > 1 ? (i / (el.categories.length - 1)) * plotW : plotW / 2);
-  const toY = (v: number) => plotY + plotH - ((v - dataMin) / range) * plotH;
-  const baseY = toY(dataMin);
-
-  ctx.save();
-  ctx.strokeStyle = '#AAAAAA'; ctx.lineWidth = 0.5; ctx.setLineDash([]);
-  ctx.beginPath(); ctx.moveTo(plotX, plotY); ctx.lineTo(plotX, plotY + plotH); ctx.lineTo(plotX + plotW, plotY + plotH); ctx.stroke();
-
-  // Draw series from back to front
-  for (let si = el.series.length - 1; si >= 0; si--) {
-    const ser = el.series[si];
-    const color = chartColor(si, ser.color);
-    ctx.globalAlpha = 0.75;
-    ctx.beginPath();
-    const pts = ser.values.map((v, i) => [toX(i), v != null ? toY(v) : baseY] as [number, number]);
-    ctx.moveTo(pts[0][0], baseY);
-    for (const [px, py] of pts) ctx.lineTo(px, py);
-    ctx.lineTo(pts[pts.length - 1][0], baseY);
-    ctx.closePath();
-    ctx.fillStyle = color; ctx.fill();
-    ctx.globalAlpha = 1.0;
-    ctx.strokeStyle = color; ctx.lineWidth = 1.5; ctx.setLineDash([]);
-    ctx.beginPath(); pts.forEach(([px, py], i) => i === 0 ? ctx.moveTo(px, py) : ctx.lineTo(px, py)); ctx.stroke();
-  }
-
-  ctx.fillStyle = '#444444'; ctx.textBaseline = 'top'; ctx.textAlign = 'center';
-  ctx.font = `${Math.max(8, h * 0.055)}px sans-serif`;
-  for (let i = 0; i < el.categories.length; i++) {
-    ctx.fillText(el.categories[i], toX(i), plotY + plotH + 2);
-  }
-  if (el.title) {
-    ctx.fillStyle = '#333333'; ctx.textAlign = 'center'; ctx.textBaseline = 'bottom';
-    ctx.font = `bold ${Math.max(9, h * 0.07)}px sans-serif`;
-    ctx.fillText(el.title, x0 + w / 2, y0 + LABEL_H * 0.4);
-  }
-  ctx.restore();
-}
-
-// ─── Pie / Doughnut chart ────────────────────────────────────────────────────
-function renderPieChart(ctx: CanvasRenderingContext2D, el: ChartElement, scale: number, isDoughnut = false) {
-  const x0 = emuToPx(el.x, scale), y0 = emuToPx(el.y, scale);
-  const w  = emuToPx(el.width, scale), h = emuToPx(el.height, scale);
-
-  const TITLE_H = el.title ? h * 0.1 : 0;
-  const LEGEND_W = w * 0.22;
-  const pieX = x0 + LEGEND_W / 2, pieY = y0 + TITLE_H;
-  const pieW = w - LEGEND_W, pieH = h - TITLE_H;
-  const cx = pieX + pieW / 2, cy = pieY + pieH / 2;
-  const r = Math.min(pieW, pieH) * 0.42;
-  const innerR = isDoughnut ? r * 0.5 : 0;
-
-  const ser = el.series[0];
-  if (!ser) return;
-
-  const vals = ser.values.map(v => v ?? 0);
-  const total = vals.reduce((a, b) => a + b, 0) || 1;
-
-  ctx.save();
-  let startAngle = -Math.PI / 2;
-  for (let i = 0; i < vals.length; i++) {
-    const sweep = (vals[i] / total) * Math.PI * 2;
-    const color = ser.dataPointColors?.[i] ?? null;
-    ctx.fillStyle = chartColor(i, color ?? ser.color);
-    ctx.beginPath();
-    ctx.moveTo(cx, cy);
-    ctx.arc(cx, cy, r, startAngle, startAngle + sweep);
-    ctx.closePath();
-    ctx.fill();
-    if (isDoughnut) {
-      // Cut out center
-      ctx.globalCompositeOperation = 'destination-out';
-      ctx.beginPath();
-      ctx.arc(cx, cy, innerR, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.globalCompositeOperation = 'source-over';
-    }
-    ctx.strokeStyle = '#ffffff'; ctx.lineWidth = 1; ctx.setLineDash([]);
-    ctx.beginPath();
-    ctx.moveTo(cx, cy);
-    ctx.arc(cx, cy, r, startAngle, startAngle + sweep);
-    ctx.closePath(); ctx.stroke();
-    startAngle += sweep;
-  }
-
-  // Legend
-  const legendX = x0 + w - LEGEND_W + 4, legendFontSz = Math.max(8, h * 0.055);
-  ctx.font = `${legendFontSz}px sans-serif`;
-  for (let i = 0; i < Math.min(vals.length, 12); i++) {
-    const lY = pieY + i * (legendFontSz + 4);
-    const color = ser.dataPointColors?.[i] ?? null;
-    ctx.fillStyle = chartColor(i, color ?? ser.color);
-    ctx.fillRect(legendX, lY + 1, legendFontSz * 0.8, legendFontSz * 0.8);
-    ctx.fillStyle = '#333333'; ctx.textBaseline = 'top'; ctx.textAlign = 'left';
-    ctx.fillText(el.categories[i] ?? '', legendX + legendFontSz + 2, lY);
-  }
-
-  if (el.title) {
-    ctx.fillStyle = '#333333'; ctx.textAlign = 'center'; ctx.textBaseline = 'top';
-    ctx.font = `bold ${Math.max(9, h * 0.08)}px sans-serif`;
-    ctx.fillText(el.title, x0 + w / 2, y0 + 4);
-  }
-  ctx.restore();
-}
-
-// ─── Scatter chart ───────────────────────────────────────────────────────────
-function renderScatterChart(ctx: CanvasRenderingContext2D, el: ChartElement, scale: number) {
-  // Scatter: we have Y values only (X positions evenly spaced); render as line without connecting
-  const x0 = emuToPx(el.x, scale), y0 = emuToPx(el.y, scale);
-  const w  = emuToPx(el.width, scale), h = emuToPx(el.height, scale);
-
-  const allVals = el.series.flatMap(s => s.values).filter((v): v is number => v != null);
-  const dataMax = allVals.length ? Math.max(...allVals) : 1;
-  const dataMin = Math.min(0, ...(allVals.length ? allVals : [0]));
-
-  const MARGIN = w * 0.07;
-  const plotX = x0 + MARGIN, plotW = w - MARGIN * 2;
-  const plotY = y0 + MARGIN * 0.5, plotH = h - MARGIN * 1.5;
-  const range = dataMax - dataMin || 1;
-  const toX = (i: number) => plotX + (el.categories.length > 1 ? (i / (el.categories.length - 1)) * plotW : plotW / 2);
-  const toY = (v: number) => plotY + plotH - ((v - dataMin) / range) * plotH;
-
-  ctx.save();
-  ctx.strokeStyle = '#AAAAAA'; ctx.lineWidth = 0.5; ctx.setLineDash([]);
-  ctx.beginPath(); ctx.moveTo(plotX, plotY); ctx.lineTo(plotX, plotY + plotH); ctx.lineTo(plotX + plotW, plotY + plotH); ctx.stroke();
-
-  for (let si = 0; si < el.series.length; si++) {
-    const ser = el.series[si];
-    ctx.fillStyle = chartColor(si, ser.color);
-    for (let i = 0; i < ser.values.length; i++) {
-      const v = ser.values[i]; if (v == null) continue;
-      ctx.beginPath(); ctx.arc(toX(i), toY(v), Math.max(3, h * 0.015), 0, Math.PI * 2); ctx.fill();
-    }
-  }
-
-  if (el.title) {
-    ctx.fillStyle = '#333333'; ctx.textAlign = 'center'; ctx.textBaseline = 'top';
-    ctx.font = `bold ${Math.max(9, h * 0.07)}px sans-serif`;
-    ctx.fillText(el.title, x0 + w / 2, y0 + 2);
-  }
-  ctx.restore();
-}
-
-function renderChart(ctx: CanvasRenderingContext2D, el: ChartElement, scale: number) {
-  if (el.chartType === 'stackedBar' || el.chartType === 'clusteredBar') {
-    renderStackedBarChart(ctx, el, scale);
-  } else if (el.chartType === 'stackedBarH' || el.chartType === 'clusteredBarH') {
-    renderStackedBarHChart(ctx, el, scale);
-  } else if (el.chartType === 'waterfall') {
-    renderWaterfallChart(ctx, el, scale);
-  } else if (el.chartType === 'line' || el.chartType === 'stackedLine' || el.chartType === 'stackedLinePct') {
-    renderLineChart(ctx, el, scale);
-  } else if (el.chartType === 'area' || el.chartType === 'stackedArea') {
-    renderAreaChart(ctx, el, scale);
-  } else if (el.chartType === 'pie') {
-    renderPieChart(ctx, el, scale, false);
-  } else if (el.chartType === 'doughnut') {
-    renderPieChart(ctx, el, scale, true);
-  } else if (el.chartType === 'scatter' || el.chartType === 'bubble') {
-    renderScatterChart(ctx, el, scale);
-  }
-}
 
 // ─── Table rendering ─────────────────────────────────────────────────────────
 
@@ -3595,7 +2885,30 @@ export async function renderSlide(
     } else if (el.type === 'table') {
       renderTable(ctx, el, scale, slideNumber, rc);
     } else if (el.type === 'chart') {
-      renderChart(ctx, el, scale);
+      renderChart(
+        ctx,
+        {
+          chartType: el.chartType,
+          title: el.title,
+          categories: el.categories,
+          series: el.series,
+          showDataLabels: el.showDataLabels,
+          valMin: el.valMin,
+          valMax: el.valMax,
+          catAxisTitle: null,
+          valAxisTitle: null,
+          catAxisHidden: el.catAxisHidden,
+          valAxisHidden: el.valAxisHidden,
+          plotAreaBg: el.plotAreaBg,
+          subtotalIndices: el.subtotalIndices,
+        },
+        {
+          x: emuToPx(el.x, scale),
+          y: emuToPx(el.y, scale),
+          w: emuToPx(el.width, scale),
+          h: emuToPx(el.height, scale),
+        },
+      );
     }
   }
